Added UPDATE and DELETE

+++ dao/src/db/user/deleteUser.js
@@ -1,6 +1,6 @@
 deleteUser = async id => {
     try {
-        const queryString = `DELETE FROM users WHERE id=${id}`
+        const queryString = `DELETE FROM jwt WHERE user_id=${id}; DELETE FROM users WHERE id=${id}`
         const result = await client.query(queryString)
         return { success: true, result: result.rowCount }
     }
